Refactor WriteTransaction God Struct
Splits the 5000-line `WriteTransaction` struct into cohesive modules.

- Moves `src/api/transaction/write_tx.rs` to `src/api/transaction/write/mod.rs`
- Extracts validation logic to `src/api/transaction/write/validation.rs`
- Extracts conflict detection to `src/api/transaction/write/conflict.rs`
- Extracts storage application logic to `src/api/transaction/write/apply.rs`
- Extracts WAL logging to `src/api/transaction/write/wal.rs`
- Extracts tests to `src/api/transaction/write/tests.rs`

This improves maintainability and cohesion by separating concerns.
The `WriteTransaction` struct fields are now `pub(crate)` to allow access from submodules.
Public API remains unchanged.
Verified with existing tests.
