+++ src/api/transaction/write/tests.rs
@@ -1,1846 +1,10 @@
-//! Write transactions with ACID guarantees.
-//!
-//! Write transactions provide full ACID properties:
-//! - **Atomicity**: All-or-nothing commit via write buffering
-//! - **Consistency**: Referential integrity validation before commit
-//! - **Isolation**: Snapshot Isolation with write-write conflict detection
-//! - **Durability**: WAL with fsync guarantees
-//!
-//! Write transactions buffer all changes in memory until commit.
-//! On commit, changes are validated and applied atomically.
+use super::*;
+use crate::api::transaction::TxIdGenerator;
+use crate::storage::wal::concurrent_system::ConcurrentWalSystemConfig;
+use tempfile::TempDir;
 
-use super::{
-    ReadOps, TransactionSnapshot, TxId, TxMetadata, TxState, TxVisibilityManager, WriteBuffer,
-    WriteOps,
-};
-use crate::core::graph::{Edge, Node};
-use crate::core::id::{EdgeId, IdGenerator, NodeId, VersionId};
-use crate::core::interning::{GLOBAL_INTERNER, InternedString};
-use crate::core::property::{PropertyMap, PropertyMapBuilder};
-use crate::core::temporal::{BiTemporalInterval, Timestamp, time};
-use crate::core::version::VersionMetadata;
-use crate::index::temporal::TemporalIndexes;
-use crate::storage::current::CurrentStorage;
-use crate::storage::historical::HistoricalStorage;
-use crate::storage::wal::concurrent_system::ConcurrentWalSystem;
-use crate::storage::wal::{DurabilityMode, WalOperation};
-use crate::utils::error::{Result, StorageError, TransactionError};
-use crate::utils::lock::{MutexExt, RwLockExt};
-use parking_lot::RwLock;
-use std::sync::{Arc, Mutex};
-
-/// Maximum allowed backward clock drift in microseconds (5 minutes).
-///
-/// If wallclock time goes backward by more than this (e.g., severe NTP adjustment),
-/// we reject the transaction to avoid timestamps jumping far into the future.
-/// 5 minutes allows for typical NTP corrections while preventing pathological cases.
-const MAX_BACKWARD_DRIFT_US: i64 = 5 * 60 * 1_000_000; // 5 minutes
-
-/// Maximum allowed forward clock jump in microseconds (1 hour).
-///
-/// If wallclock time jumps forward by more than this (e.g., system clock set manually),
-/// we reject the transaction to avoid creating timestamps in the far future that would
-/// break temporal query semantics.
-const MAX_FORWARD_JUMP_US: i64 = 60 * 60 * 1_000_000; // 1 hour
-
-/// Maximum offset for valid_from timestamps in the future.
-///
-/// Users can backdate facts (valid_from < transaction_time) for historical corrections,
-/// but we limit how far into the future they can set valid_from to prevent abuse and
-/// maintain query semantics.
-const MAX_VALID_TIME_FUTURE_OFFSET_US: i64 = 365 * 24 * 60 * 60 * 1_000_000; // 1 year
-
-/// Write transaction with full ACID guarantees.
-///
-/// Write transactions buffer all operations in memory and apply them
-/// atomically on commit. This ensures consistency and enables rollback.
-///
-/// # Example
-///
-/// ```ignore
-/// let mut tx = db.write_transaction();
-/// let node_id = tx.create_node("Person", props)?;
-/// tx.create_edge(node_id, other, "KNOWS", edge_props)?;
-/// tx.commit()?;  // or tx.rollback()
-/// ```
-pub struct WriteTransaction {
-    tx_id: TxId,
-    start_timestamp: Timestamp,
-    state: TxState,
-
-    // Snapshot for Snapshot Isolation
-    snapshot: TransactionSnapshot,
-
-    // Write buffer for uncommitted changes
-    buffer: WriteBuffer,
-
-    // Shared references to storage (Arc for zero-copy sharing)
-    current: Arc<CurrentStorage>,
-    historical: Arc<RwLock<HistoricalStorage>>,
-    temporal_indexes: Arc<TemporalIndexes>,
-    wal: Arc<ConcurrentWalSystem>,
-    current_timestamp: Arc<Mutex<Timestamp>>,
-    visibility_manager: Arc<TxVisibilityManager>,
-
-    // ID generators (needed for creating new entities)
-    node_id_gen: Arc<Mutex<IdGenerator>>,
-    edge_id_gen: Arc<Mutex<IdGenerator>>,
-    version_id_gen: Arc<Mutex<IdGenerator>>,
-
-    /// Durability mode for this transaction's commit
-    durability_mode: DurabilityMode,
-}
-
-impl WriteTransaction {
-    /// Create a new write transaction with default durability mode (Synchronous).
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn new(
-        tx_id: TxId,
-        snapshot: TransactionSnapshot,
-        current: Arc<CurrentStorage>,
-        historical: Arc<RwLock<HistoricalStorage>>,
-        temporal_indexes: Arc<TemporalIndexes>,
-        wal: Arc<ConcurrentWalSystem>,
-        current_timestamp: Arc<Mutex<Timestamp>>,
-        visibility_manager: Arc<TxVisibilityManager>,
-        node_id_gen: Arc<Mutex<IdGenerator>>,
-        edge_id_gen: Arc<Mutex<IdGenerator>>,
-        version_id_gen: Arc<Mutex<IdGenerator>>,
-    ) -> Self {
-        Self::new_with_durability(
-            tx_id,
-            snapshot,
-            current,
-            historical,
-            temporal_indexes,
-            wal,
-            current_timestamp,
-            visibility_manager,
-            node_id_gen,
-            edge_id_gen,
-            version_id_gen,
-            DurabilityMode::Synchronous,
-        )
-    }
-
-    /// Create a new write transaction with a specific durability mode.
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn new_with_durability(
-        tx_id: TxId,
-        snapshot: TransactionSnapshot,
-        current: Arc<CurrentStorage>,
-        historical: Arc<RwLock<HistoricalStorage>>,
-        temporal_indexes: Arc<TemporalIndexes>,
-        wal: Arc<ConcurrentWalSystem>,
-        current_timestamp: Arc<Mutex<Timestamp>>,
-        visibility_manager: Arc<TxVisibilityManager>,
-        node_id_gen: Arc<Mutex<IdGenerator>>,
-        edge_id_gen: Arc<Mutex<IdGenerator>>,
-        version_id_gen: Arc<Mutex<IdGenerator>>,
-        durability_mode: DurabilityMode,
-    ) -> Self {
-        WriteTransaction {
-            tx_id,
-            start_timestamp: time::now(),
-            state: TxState::Active,
-            snapshot,
-            buffer: WriteBuffer::new(),
-            current,
-            historical,
-            temporal_indexes,
-            wal,
-            current_timestamp,
-            visibility_manager,
-            node_id_gen,
-            edge_id_gen,
-            version_id_gen,
-            durability_mode,
-        }
-    }
-
-    /// Get transaction metadata.
-    pub fn metadata(&self) -> TxMetadata {
-        TxMetadata {
-            tx_id: self.tx_id,
-            start_timestamp: self.start_timestamp,
-            commit_timestamp: None,
-            state: self.state,
-            is_read_only: false,
-        }
-    }
-
-    /// Get transaction ID.
-    pub fn tx_id(&self) -> TxId {
-        self.tx_id
-    }
-
-    /// Commit the transaction.
-    ///
-    /// This validates all buffered writes and applies them atomically
-    /// to the storage. If validation fails or any operation fails,
-    /// the transaction is rolled back.
-    ///
-    /// The durability behavior depends on the transaction's durability mode:
-    /// - **Synchronous**: Waits for fsync (maximum durability, default)
-    /// - **Async**: Returns after flush to OS cache (background thread syncs)
-    /// - **GroupCommit**: Waits for batch fsync (ACID + high throughput)
-    /// - **AsyncBatched**: Returns after flush to OS cache, batched fsync in background (<100µs latency)
-    pub fn commit(self) -> Result<()> {
-        self.commit_with_timestamp().map(|_| ())
-    }
-
-    /// Commit the transaction and return the commit timestamp.
-    ///
-    /// This is useful for benchmarks and tests that need to query the database
-    /// at the exact commit timestamp to verify temporal semantics.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let mut tx = db.write_transaction()?;
-    /// let node_id = tx.create_node("Person", properties)?;
-    /// let commit_ts = tx.commit_with_timestamp()?;
-    ///
-    /// // Query at exact commit timestamp
-    /// let node = db.get_node_at_time(node_id, commit_ts, commit_ts)?;
-    /// ```
-    ///
-    /// # Durability Modes
-    ///
-    /// - **Synchronous**: WAL fsynced to disk before returning (ACID, ~1.5ms)
-    /// - **Async**: Returns after flush to OS cache (background thread syncs)
-    /// - **GroupCommit**: Waits for batch fsync (ACID + high throughput)
-    /// - **AsyncBatched**: Returns after flush to OS cache, batched fsync in background (<100µs latency)
-    pub fn commit_with_timestamp(mut self) -> Result<Timestamp> {
-        #[cfg(feature = "observability")]
-        let _span = tracing::info_span!(
-            "transaction_commit",
-            tx_id = %self.tx_id
-        )
-        .entered();
-
-        #[cfg(feature = "observability")]
-        let commit_start = std::time::Instant::now();
-
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Transition to preparing state
-        self.state = TxState::Preparing;
-
-        // Validate all buffered writes
-        self.validate()?;
-
-        // Detect write-write conflicts (Snapshot Isolation)
-        self.detect_conflicts()?;
-
-        // Acquire commit timestamp and perform mode-aware WAL flush.
-        //
-        // CRITICAL: We must hold the timestamp lock until WAL logging is complete
-        // to prevent a race condition where transactions commit out-of-order.
-        //
-        // Timestamp management for Bi-Temporal Database:
-        //
-        // For temporal queries to work correctly, we MUST use wallclock timestamps
-        // for transaction_time, not a logical clock. This allows querying historical
-        // state at specific points in time (e.g., "what was the state at 2PM yesterday?").
-        //
-        // Monotonicity: Wallclock time is monotonically increasing (assuming NTP is working),
-        // which satisfies the ordering requirements for Snapshot Isolation:
-        // - commit_ts > snapshot_ts for transactions that started before this commit
-        // - future snapshots will have timestamp >= this commit
-        //
-        // DURABILITY MODES (handled by ConcurrentWalSystem):
-        // - Synchronous: Appends drain and flush immediately with fsync
-        // - Async: Appends go to ring buffers, background thread syncs
-        // - GroupCommit: Appends go to ring buffers, wait for epoch completion
-        let commit_timestamp = {
-            #[cfg(feature = "observability")]
-            let ts_lock_start = std::time::Instant::now();
-
-            let mut ts = self.current_timestamp.lock_or_err().map_err(|e| {
-                TransactionError::CommitFailed {
-                    reason: format!("timestamp lock poisoned: {}", e),
-                }
-            })?;
-
-            #[cfg(feature = "observability")]
-            let ts_lock_acquired = std::time::Instant::now();
-
-            // Phase 2: Use HLC for distributed temporal consistency
-            // Get current physical wallclock
-            let current_wallclock = crate::core::temporal::time::now();
-
-            // Check for pathological clock skew using wallclock components
-            let drift = current_wallclock.wallclock() - ts.wallclock();
-
-            // Backward drift check: prevent timestamps jumping far into future
-            if drift < -MAX_BACKWARD_DRIFT_US {
-                return Err(TransactionError::ClockSkew {
-                    wallclock: current_wallclock.wallclock(),
-                    previous: ts.wallclock(),
-                    drift_us: drift,
-                    max_allowed: -MAX_BACKWARD_DRIFT_US,
-                }
-                .into());
-            }
-
-            // Forward jump check: prevent timestamps in far future
-            if drift > MAX_FORWARD_JUMP_US {
-                return Err(TransactionError::ClockSkew {
-                    wallclock: current_wallclock.wallclock(),
-                    previous: ts.wallclock(),
-                    drift_us: drift,
-                    max_allowed: MAX_FORWARD_JUMP_US,
-                }
-                .into());
-            }
-
-            // Phase 2: Use HLC .send() method for monotonic timestamp generation
-            // This ensures: if wallclock advances, reset logical; otherwise increment logical
-            let commit = ts.send(current_wallclock.wallclock()).map_err(|e| {
-                TransactionError::CommitFailed {
-                    reason: format!("HLC timestamp generation failed: {}", e),
-                }
-            })?;
-
-            // Observability: Warn about clock skew issues
-            #[cfg(feature = "observability")]
-            {
-                // Clock went backwards: wallclock < previous wallclock
-                if current_wallclock.wallclock() < ts.wallclock() {
-                    tracing::warn!(
-                        wallclock_ts = %current_wallclock,
-                        prev_ts = %ts,
-                        skew_us = ts.wallclock() - current_wallclock.wallclock(),
-                        logical_counter = commit.logical(),
-                        "Clock skew detected: wallclock went backwards (NTP adjustment?)"
-                    );
-                } else if commit.wallclock() > ts.wallclock() + 60_000_000 {
-                    // Large forward jump (>60 seconds)
-                    tracing::warn!(
-                        wallclock_ts = %current_wallclock,
-                        prev_ts = %ts,
-                        jump_us = commit.wallclock() - ts.wallclock(),
-                        "Large clock jump detected: timestamps will be lumpy"
-                    );
-                }
-            }
-
-            // Update current_timestamp for next transaction's snapshot
-            *ts = commit;
-
-            #[cfg(feature = "observability")]
-            let wal_start = std::time::Instant::now();
-
-            // Log operations to WAL (lock-free striped append!)
-            // This must happen BEFORE applying changes for durability.
-            self.log_operations_to_wal(commit)?;
-
-            #[cfg(feature = "observability")]
-            let wal_logged = std::time::Instant::now();
-
-            // Commit with configured durability mode
-            // For Sync: drains and flushes immediately
-            // For Async: returns immediately
-            // For GroupCommit: registers and returns epoch
-            let wait_epoch = self.wal.commit()?;
-
-            #[cfg(feature = "observability")]
-            let wal_commit_completed = std::time::Instant::now();
-
-            // For GroupCommit mode, wait for the epoch to be flushed.
-            // AsyncBatched mode returns an epoch but does NOT wait.
-            if let Some(epoch) = wait_epoch
-                && let Some(gc) = self.wal.group_commit_coordinator()
-                && self.durability_mode.waits_for_durability()
-            {
-                gc.wait_for_flush(epoch)?;
-            }
-
-            #[cfg(feature = "observability")]
-            {
-                // Record detailed breakdown for Honeycomb
-                let ts_lock_wait_us =
-                    ts_lock_acquired.duration_since(ts_lock_start).as_micros() as u64;
-                let wal_log_us = wal_logged.duration_since(wal_start).as_micros() as u64;
-                let wal_commit_us =
-                    wal_commit_completed.duration_since(wal_logged).as_micros() as u64;
-                let total_us = wal_commit_completed
-                    .duration_since(ts_lock_start)
-                    .as_micros() as u64;
-
-                // Calculate total commit duration for Honeycomb queries
-                let total_commit_us = commit_start.elapsed().as_micros() as u64;
-                let operations_count = self.buffer.operations().len();
-
-                tracing::info!(
-                    ts_lock_wait_us,
-                    wal_log_us,
-                    wal_commit_us,
-                    total_us,
-                    total_commit_us,
-                    operations_count,
-                    commit_ts = %commit,
-                    durability_mode = ?self.durability_mode,
-                    "Transaction commit breakdown (concurrent WAL)"
-                );
-            }
-
-            commit
-        };
-
-        // Apply all changes atomically
-        self.apply_changes(commit_timestamp)?;
-
-        // Notify temporal vector index of transaction completion (for snapshot creation)
-        // Only call this if the transaction modified vector properties to avoid unnecessary overhead
-        if self.buffer.has_vector_operations() {
-            self.current.on_temporal_vector_transaction()?;
-        }
-
-        // Register commit with visibility manager
-        self.visibility_manager
-            .register_commit(self.tx_id, commit_timestamp);
-
-        // Mark as committed
-        self.state = TxState::Committed;
-
-        #[cfg(feature = "observability")]
-        {
-            let total_commit_us = commit_start.elapsed().as_micros() as u64;
-            tracing::debug!(
-                total_commit_us,
-                tx_id = %self.tx_id,
-                "Transaction committed"
-            );
-        }
-
-        Ok(commit_timestamp)
-    }
-
-    /// Rollback the transaction.
-    ///
-    /// Discards all buffered writes. This is automatically called
-    /// if the transaction is dropped without committing.
-    pub fn rollback(mut self) -> Result<()> {
-        if self.state == TxState::Committed {
-            return Err(TransactionError::AlreadyCommitted {
-                tx_id: self.tx_id.as_u64(),
-            }
-            .into());
-        }
-
-        // Clear the write buffer
-        self.buffer.clear();
-
-        // Register abort with visibility manager
-        self.visibility_manager.register_abort(self.tx_id);
-
-        self.state = TxState::Aborted;
-
-        Ok(())
-    }
-
-    /// Check if this transaction has any node writes (create, update, delete).
-    pub(crate) fn has_node_writes(&self) -> bool {
-        self.buffer.operations().iter().any(|op| {
-            matches!(
-                op,
-                super::BufferedWrite::CreateNode { .. }
-                    | super::BufferedWrite::UpdateNode { .. }
-                    | super::BufferedWrite::DeleteNode { .. }
-            )
-        })
-    }
-
-    /// Check if this transaction has any edge writes (create, update, delete).
-    pub(crate) fn has_edge_writes(&self) -> bool {
-        self.buffer.operations().iter().any(|op| {
-            matches!(
-                op,
-                super::BufferedWrite::CreateEdge { .. }
-                    | super::BufferedWrite::UpdateEdge { .. }
-                    | super::BufferedWrite::DeleteEdge { .. }
-            )
-        })
-    }
-
-    /// Check if this transaction has any vector property writes.
-    pub(crate) fn has_vector_writes(&self) -> bool {
-        self.buffer.has_vector_operations()
-    }
-
-    // =========================================================================
-    // Bi-Temporal Validation (Phase 5)
-    // =========================================================================
-
-    /// Validate that valid_from is not too far in the future.
-    ///
-    /// Limits how far users can set valid_from into the future to prevent
-    /// abuse and maintain temporal query semantics.
-    fn validate_valid_from_future(&self, valid_from: Timestamp) -> Result<()> {
-        let current = time::now();
-        let future_offset = valid_from.wallclock() as i128 - current.wallclock() as i128;
-
-        if future_offset > MAX_VALID_TIME_FUTURE_OFFSET_US as i128 {
-            return Err(
-                crate::utils::error::TemporalError::ValidTimeTooFarInFuture {
-                    valid_from,
-                    current_time: current,
-                    max_future_offset_us: MAX_VALID_TIME_FUTURE_OFFSET_US,
-                }
-                .into(),
-            );
-        }
-        Ok(())
-    }
-
-    /// Validate that valid_from is not before entity creation for updates/deletes.
-    ///
-    /// For updates and deletes, valid_from must be >= the entity's original
-    /// creation time to maintain temporal consistency.
-    fn validate_valid_from_not_before_creation(
-        &self,
-        entity_id: &str,
-        entity_creation_time: Timestamp,
-        valid_from: Timestamp,
-    ) -> Result<()> {
-        if valid_from < entity_creation_time {
-            return Err(
-                crate::utils::error::TemporalError::ValidTimeBeforeEntityCreation {
-                    valid_from,
-                    entity_creation_time,
-                    entity_id: entity_id.to_string(),
-                }
-                .into(),
-            );
-        }
-        Ok(())
-    }
-
-    /// Validate all buffered writes.
-    ///
-    /// Checks:
-    /// - Referential integrity (edges reference valid nodes)
-    /// - No constraint violations
-    fn validate(&self) -> Result<()> {
-        for write in self.buffer.operations() {
-            match write {
-                super::BufferedWrite::CreateEdge { source, target, .. }
-                | super::BufferedWrite::UpdateEdge { source, target, .. } => {
-                    // Check that source and target nodes exist
-                    // They might exist in current storage or be created in this transaction
-                    if !self.buffer.has_modified_node(*source)
-                        && self.current.get_node(*source).is_err()
-                    {
-                        return Err(TransactionError::ValidationFailed {
-                            reason: format!("Edge source node {:?} does not exist", source),
-                        }
-                        .into());
-                    }
-                    if !self.buffer.has_modified_node(*target)
-                        && self.current.get_node(*target).is_err()
-                    {
-                        return Err(TransactionError::ValidationFailed {
-                            reason: format!("Edge target node {:?} does not exist", target),
-                        }
-                        .into());
-                    }
-                }
-                _ => {
-                    // Other operations don't need validation
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Detect write-write conflicts for Snapshot Isolation.
-    ///
-    /// Checks if any entity modified by this transaction has been committed
-    /// by another transaction after our snapshot was taken. This implements
-    /// the First-Committer-Wins rule of Snapshot Isolation.
-    ///
-    /// # Errors
-    ///
-    /// Returns `SerializationFailure` if a write-write conflict is detected.
-    fn detect_conflicts(&self) -> Result<()> {
-        for write in self.buffer.operations() {
-            match write {
-                // UpdateNode: check if node was modified or deleted after our snapshot
-                super::BufferedWrite::UpdateNode { node_id, .. } => {
-                    match self.current.get_node(*node_id) {
-                        Ok(current_node) => {
-                            // Node exists - check if it was modified after our snapshot
-                            if let Some(commit_ts) = current_node.metadata.commit_timestamp
-                                && commit_ts > self.snapshot.snapshot_timestamp
-                            {
-                                return Err(TransactionError::SerializationFailure {
-                                    entity: format!("{:?}", node_id),
-                                    reason: format!(
-                                        "Version committed at {} after snapshot at {}",
-                                        commit_ts, self.snapshot.snapshot_timestamp
-                                    ),
-                                }
-                                .into());
-                            }
-                        }
-                        Err(_) => {
-                            // Node doesn't exist in current storage. Since we successfully
-                            // called update_node() earlier (which reads from current storage),
-                            // the node must have been deleted by another transaction.
-                            return Err(TransactionError::SerializationFailure {
-                                entity: format!("{:?}", node_id),
-                                reason: "Node was deleted by another transaction".to_string(),
-                            }
-                            .into());
-                        }
-                    }
-                }
-
-                // UpdateEdge: check if edge was modified or deleted after our snapshot
-                super::BufferedWrite::UpdateEdge { edge_id, .. } => {
-                    match self.current.get_edge(*edge_id) {
-                        Ok(current_edge) => {
-                            // Edge exists - check if it was modified after our snapshot
-                            if let Some(commit_ts) = current_edge.metadata.commit_timestamp
-                                && commit_ts > self.snapshot.snapshot_timestamp
-                            {
-                                return Err(TransactionError::SerializationFailure {
-                                    entity: format!("{:?}", edge_id),
-                                    reason: format!(
-                                        "Version committed at {} after snapshot at {}",
-                                        commit_ts, self.snapshot.snapshot_timestamp
-                                    ),
-                                }
-                                .into());
-                            }
-                        }
-                        Err(_) => {
-                            // Edge doesn't exist - it was deleted by another transaction
-                            return Err(TransactionError::SerializationFailure {
-                                entity: format!("{:?}", edge_id),
-                                reason: "Edge was deleted by another transaction".to_string(),
-                            }
-                            .into());
-                        }
-                    }
-                }
-
-                // DeleteNode: check if node was modified after our snapshot
-                super::BufferedWrite::DeleteNode { node_id, .. } => {
-                    // Get current version from storage
-                    if let Ok(current_node) = self.current.get_node(*node_id)
-                        && let Some(commit_ts) = current_node.metadata.commit_timestamp
-                        && commit_ts > self.snapshot.snapshot_timestamp
-                    {
-                        return Err(TransactionError::SerializationFailure {
-                            entity: format!("{:?}", node_id),
-                            reason: format!(
-                                "Version committed at {} after snapshot at {}",
-                                commit_ts, self.snapshot.snapshot_timestamp
-                            ),
-                        }
-                        .into());
-                    }
-                }
-
-                // DeleteEdge: check if edge was modified after our snapshot
-                super::BufferedWrite::DeleteEdge { edge_id, .. } => {
-                    // Get current version from storage
-                    if let Ok(current_edge) = self.current.get_edge(*edge_id)
-                        && let Some(commit_ts) = current_edge.metadata.commit_timestamp
-                        && commit_ts > self.snapshot.snapshot_timestamp
-                    {
-                        return Err(TransactionError::SerializationFailure {
-                            entity: format!("{:?}", edge_id),
-                            reason: format!(
-                                "Version committed at {} after snapshot at {}",
-                                commit_ts, self.snapshot.snapshot_timestamp
-                            ),
-                        }
-                        .into());
-                    }
-                }
-
-                // CreateNode and CreateEdge don't need conflict detection
-                // since they're creating new entities that didn't exist before
-                _ => {}
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Log all buffered operations to WAL.
-    ///
-    /// This ensures durability - operations are logged before being applied.
-    /// Uses lock-free appends to the concurrent WAL system.
-    fn log_operations_to_wal(&self, _commit_timestamp: Timestamp) -> Result<()> {
-        for write in self.buffer.operations() {
-            let operation = match write {
-                super::BufferedWrite::CreateNode {
-                    node_id,
-                    label,
-                    properties,
-                    valid_from,
-                    ..
-                } => {
-                    // No allocation! Just copy the 4-byte InternedString ID
-                    WalOperation::CreateNode {
-                        node_id: *node_id,
-                        label: *label,
-                        properties: properties.clone(),
-                        valid_from: *valid_from,
-                    }
-                }
-                super::BufferedWrite::CreateEdge {
-                    edge_id,
-                    source,
-                    target,
-                    label,
-                    properties,
-                    valid_from,
-                    ..
-                } => {
-                    // No allocation! Just copy the 4-byte InternedString ID
-                    WalOperation::CreateEdge {
-                        edge_id: *edge_id,
-                        source: *source,
-                        target: *target,
-                        label: *label,
-                        properties: properties.clone(),
-                        valid_from: *valid_from,
-                    }
-                }
-                super::BufferedWrite::UpdateNode {
-                    node_id,
-                    version_id,
-                    label,
-                    properties,
-                    valid_from,
-                    ..
-                } => {
-                    // No allocation! Just copy the 4-byte InternedString ID
-                    WalOperation::UpdateNode {
-                        node_id: *node_id,
-                        version_id: *version_id,
-                        label: *label,
-                        properties: properties.clone(),
-                        valid_from: *valid_from,
-                    }
-                }
-                super::BufferedWrite::UpdateEdge {
-                    edge_id,
-                    version_id,
-                    label,
-                    properties,
-                    valid_from,
-                    ..
-                } => {
-                    // No allocation! Just copy the 4-byte InternedString ID
-                    WalOperation::UpdateEdge {
-                        edge_id: *edge_id,
-                        version_id: *version_id,
-                        label: *label,
-                        properties: properties.clone(),
-                        valid_from: *valid_from,
-                    }
-                }
-                super::BufferedWrite::DeleteNode {
-                    node_id,
-                    valid_from,
-                } => WalOperation::DeleteNode {
-                    node_id: *node_id,
-                    valid_from: *valid_from,
-                },
-                super::BufferedWrite::DeleteEdge {
-                    edge_id,
-                    valid_from,
-                } => WalOperation::DeleteEdge {
-                    edge_id: *edge_id,
-                    valid_from: *valid_from,
-                },
-            };
-
-            // Append to WAL (lock-free!)
-            self.wal.append_async(operation)?;
-        }
-
-        Ok(())
-    }
-
-    /// Apply all buffered changes to storage.
-    ///
-    /// # Lock Consolidation Optimization
-    ///
-    /// Acquires locks for historical storage and temporal indexes once before
-    /// processing all operations, reducing lock overhead from O(N) to O(1).
-    ///
-    /// ## Performance Trade-offs
-    ///
-    /// **Benefits:**
-    /// - Reduces lock acquisitions from 2N (per operation) to 2 (total)
-    /// - For 1000 operations: 2000 lock ops → 2 lock ops
-    ///
-    /// **Costs:**
-    /// - Longer critical section (holds locks for entire operation batch)
-    /// - May reduce concurrency for other transactions during this time
-    /// - Acceptable for most workloads; monitor if contention becomes an issue
-    ///
-    /// ## Error Handling
-    ///
-    /// If any operation fails, locks are released via RAII (automatic drop).
-    /// Note: Partial application is NOT rolled back here since this runs AFTER
-    /// WAL flush (see commit() at line 203). WAL recovery will handle consistency.
-    ///
-    /// ## Lock Ordering Hierarchy
-    ///
-    /// To prevent deadlocks, locks must be acquired in this order:
-    /// 1. `historical` - historical storage (acquired first)
-    /// 2. `temporal_indexes` - temporal indexes (acquired second)
-    /// 3. `version_id_gen` - version ID generator (acquired later for tombstones)
-    ///
-    /// ## Helper Functions
-    ///
-    /// Helper function to create a bi-temporal interval with proper closing logic.
-    ///
-    /// This centralizes the interval construction logic to ensure both historical storage
-    /// and temporal indexes use exactly the same interval semantics.
-    ///
-    /// # Arguments
-    ///
-    /// * `valid_from` - When the fact became valid in reality (user-controlled)
-    /// * `tx_time` - When the fact was recorded in the database (system-controlled)
-    /// * `is_tombstone` - Whether this is a deletion tombstone (closes valid_time immediately)
-    ///
-    /// # Returns
-    ///
-    /// A properly constructed `BiTemporalInterval`:
-    /// - Regular versions: Open-ended valid_time `[valid_from, ∞)`
-    /// - Tombstones: Closed valid_time `[valid_from, valid_from)` (empty interval)
-    #[inline]
-    fn create_temporal_interval(
-        valid_from: Timestamp,
-        tx_time: Timestamp,
-        is_tombstone: bool,
-    ) -> BiTemporalInterval {
-        let mut temporal = BiTemporalInterval::with_valid_time(valid_from, tx_time);
-        if is_tombstone {
-            // Close valid_time at valid_from to create empty interval [valid_from, valid_from)
-            // This represents "entity is no longer valid starting from this point"
-            temporal = temporal.close_valid_time(valid_from);
-        }
-        temporal
-    }
-
-    /// Helper function to apply node writes (both create and update operations).
-    ///
-    /// # Arguments
-    ///
-    /// * `is_create` - True for CreateNode, false for UpdateNode
-    /// * `node_id` - The ID of the node
-    /// * `version_id` - The version ID for this write
-    /// * `label` - The node label
-    /// * `properties` - The node properties
-    /// * `valid_from` - When the fact became valid in reality
-    /// * `commit_timestamp` - The commit timestamp (transaction time)
-    /// * `historical` - Mutable reference to historical storage
-    #[allow(clippy::too_many_arguments)]
-    fn apply_node_write(
-        &self,
-        is_create: bool,
-        node_id: NodeId,
-        version_id: VersionId,
-        label: InternedString,
-        properties: PropertyMap,
-        valid_from: Timestamp,
-        commit_timestamp: Timestamp,
-        historical: &mut HistoricalStorage,
-    ) -> Result<()> {
-        // Create node with proper transaction metadata
-        let metadata = VersionMetadata::new(self.tx_id, commit_timestamp);
-        let node = Node::with_metadata(node_id, label, properties.clone(), version_id, metadata);
-
-        // Insert or update in current storage
-        if is_create {
-            self.current.insert_node_direct(node, commit_timestamp)?;
-        } else {
-            self.current.update_node_direct(node, commit_timestamp)?;
-
-            // For updates, explicitly close the current version's transaction_time in historical storage.
-            //
-            // WHY THIS IS NEEDED:
-            // When multiple updates to the same node occur in a SINGLE transaction, they all share
-            // the same commit_timestamp. The auto-closing logic in add_node_version() checks:
-            //   if new_tx_time > prev_tx_time { close prev_version }
-            // But when new_tx_time == prev_tx_time (same transaction), this check fails.
-            //
-            // EXAMPLE SCENARIO:
-            //   db.write(|tx| {
-            //       tx.update_node(node_id, props1)?;  // First update
-            //       tx.update_node(node_id, props2)?;  // Second update - SAME commit_timestamp!
-            //       Ok(())
-            //   })?;
-            //
-            // Without explicit closing, the version chain would be corrupt:
-            // - Version 1: tx_time [T1, ∞)  <- Should be [T1, T2)
-            // - Version 2: tx_time [T2, ∞)  <- Never created because T2 == T2
-            //
-            // With explicit closing:
-            // - Version 1: tx_time [T1, T2)  <- Explicitly closed before adding Version 2
-            // - Version 2: tx_time [T2, ∞)  <- Correctly added with same T2
-            //
-            // See test: test_multiple_updates_same_transaction in tests/benchmark_validation.rs
-            if let Some(current_version_id) = historical.get_current_node_version(node_id) {
-                historical
-                    .close_node_version_transaction_time(current_version_id, commit_timestamp)?;
-            }
-        }
-
-        // Store in historical storage (consume properties, avoiding second clone)
-        historical.add_node_version(
-            node_id,
-            version_id,
-            valid_from,
-            commit_timestamp,
-            label,
-            properties,
-            false, // not a tombstone
-        )?;
-
-        // Index in temporal indexes with bi-temporal interval
-        // Use the same interval construction logic as historical storage
-        let temporal = Self::create_temporal_interval(valid_from, commit_timestamp, false);
-        self.temporal_indexes
-            .insert_node_version(node_id, version_id, temporal)?;
-
-        Ok(())
-    }
-
-    /// Helper function to apply edge writes (both create and update operations).
-    ///
-    /// # Arguments
-    ///
-    /// * `is_create` - True for CreateEdge, false for UpdateEdge
-    /// * `edge_id` - The ID of the edge
-    /// * `version_id` - The version ID for this write
-    /// * `source` - The source node ID
-    /// * `target` - The target node ID
-    /// * `label` - The edge label
-    /// * `properties` - The edge properties
-    /// * `valid_from` - When the fact became valid in reality
-    /// * `commit_timestamp` - The commit timestamp (transaction time)
-    /// * `historical` - Mutable reference to historical storage
-    #[allow(clippy::too_many_arguments)]
-    fn apply_edge_write(
-        &self,
-        is_create: bool,
-        edge_id: EdgeId,
-        version_id: VersionId,
-        source: NodeId,
-        target: NodeId,
-        label: InternedString,
-        properties: PropertyMap,
-        valid_from: Timestamp,
-        commit_timestamp: Timestamp,
-        historical: &mut HistoricalStorage,
-    ) -> Result<()> {
-        // Create edge with proper transaction metadata
-        let metadata = VersionMetadata::new(self.tx_id, commit_timestamp);
-        let edge = Edge::with_metadata(
-            edge_id,
-            label,
-            source,
-            target,
-            properties.clone(),
-            version_id,
-            metadata,
-        );
-
-        // Insert or update in current storage
-        if is_create {
-            self.current.insert_edge_direct(edge)?;
-        } else {
-            self.current.update_edge_direct(edge)?;
-
-            // For updates, explicitly close the current version's transaction_time in historical storage.
-            //
-            // WHY THIS IS NEEDED:
-            // When multiple updates to the same edge occur in a SINGLE transaction, they all share
-            // the same commit_timestamp. The auto-closing logic in add_edge_version() checks:
-            //   if new_tx_time > prev_tx_time { close prev_version }
-            // But when new_tx_time == prev_tx_time (same transaction), this check fails.
-            //
-            // EXAMPLE SCENARIO:
-            //   db.write(|tx| {
-            //       tx.update_edge(edge_id, props1)?;  // First update
-            //       tx.update_edge(edge_id, props2)?;  // Second update - SAME commit_timestamp!
-            //       Ok(())
-            //   })?;
-            //
-            // Without explicit closing, the version chain would be corrupt:
-            // - Version 1: tx_time [T1, ∞)  <- Should be [T1, T2)
-            // - Version 2: tx_time [T2, ∞)  <- Never created because T2 == T2
-            //
-            // With explicit closing:
-            // - Version 1: tx_time [T1, T2)  <- Explicitly closed before adding Version 2
-            // - Version 2: tx_time [T2, ∞)  <- Correctly added with same T2
-            //
-            // See test: test_multiple_updates_same_transaction in tests/benchmark_validation.rs
-            if let Some(current_version_id) = historical.get_current_edge_version(edge_id) {
-                historical
-                    .close_edge_version_transaction_time(current_version_id, commit_timestamp)?;
-            }
-        }
-
-        // Store in historical storage (consume properties, avoiding second clone)
-        historical.add_edge_version(
-            edge_id,
-            version_id,
-            valid_from,
-            commit_timestamp,
-            label,
-            source,
-            target,
-            properties,
-            false, // not a tombstone
-        )?;
-
-        // Index in temporal indexes with bi-temporal interval
-        // Use the same interval construction logic as historical storage
-        let temporal = Self::create_temporal_interval(valid_from, commit_timestamp, false);
-        self.temporal_indexes
-            .insert_edge_version(edge_id, version_id, temporal)?;
-
-        Ok(())
-    }
-
-    /// Helper function to apply node deletion.
-    ///
-    /// # Arguments
-    ///
-    /// * `node_id` - The ID of the node to delete
-    /// * `commit_timestamp` - The commit timestamp
-    /// * `tombstone_id` - The version ID for the tombstone
-    /// * `historical` - Mutable reference to historical storage
-    fn apply_node_delete(
-        &self,
-        node_id: NodeId,
-        valid_from: Timestamp,
-        commit_timestamp: Timestamp,
-        tombstone_id: VersionId,
-        historical: &mut HistoricalStorage,
-    ) -> Result<()> {
-        // Get the node before deleting
-        let node = self.current.get_node(node_id)?;
-
-        // Close the current version's transaction_time in historical storage
-        // This marks the end of this version's visibility
-        if let Some(current_version_id) = historical.get_current_node_version(node_id) {
-            historical.close_node_version_transaction_time(current_version_id, commit_timestamp)?;
-        }
-
-        // Create tombstone interval using centralized logic
-        // Tombstones have closed valid_time intervals [tx_time, tx_time) (empty interval)
-        let tombstone_temporal = Self::create_temporal_interval(valid_from, commit_timestamp, true);
-
-        // Add tombstone version to historical storage
-        // The is_tombstone=true flag ensures add_node_version uses the same interval logic
-        historical.add_node_version(
-            node_id,
-            tombstone_id,
-            valid_from,
-            commit_timestamp,
-            node.label,
-            node.properties.clone(),
-            true, // is_tombstone
-        )?;
-
-        // Index the tombstone version with the same interval
-        self.temporal_indexes
-            .insert_node_version(node_id, tombstone_id, tombstone_temporal)?;
-
-        // Delete from current storage
-        self.current.delete_node_direct(node_id, commit_timestamp)?;
-
-        Ok(())
-    }
-
-    /// Helper function to apply edge deletion.
-    ///
-    /// # Arguments
-    ///
-    /// * `edge_id` - The ID of the edge to delete
-    /// * `commit_timestamp` - The commit timestamp
-    /// * `tombstone_id` - The version ID for the tombstone
-    /// * `historical` - Mutable reference to historical storage
-    fn apply_edge_delete(
-        &self,
-        edge_id: EdgeId,
-        valid_from: Timestamp,
-        commit_timestamp: Timestamp,
-        tombstone_id: VersionId,
-        historical: &mut HistoricalStorage,
-    ) -> Result<()> {
-        // Get the edge before deleting
-        let edge = self.current.get_edge(edge_id)?;
-
-        // Close the current version's transaction_time in historical storage
-        // This marks the end of this version's visibility
-        if let Some(current_version_id) = historical.get_current_edge_version(edge_id) {
-            historical.close_edge_version_transaction_time(current_version_id, commit_timestamp)?;
-        }
-
-        // Create tombstone interval using centralized logic
-        // Tombstones have closed valid_time intervals [tx_time, tx_time) (empty interval)
-        let tombstone_temporal = Self::create_temporal_interval(valid_from, commit_timestamp, true);
-
-        // Add tombstone version to historical storage
-        // The is_tombstone=true flag ensures add_edge_version uses the same interval logic
-        historical.add_edge_version(
-            edge_id,
-            tombstone_id,
-            valid_from,
-            commit_timestamp,
-            edge.label,
-            edge.source,
-            edge.target,
-            edge.properties.clone(),
-            true, // is_tombstone
-        )?;
-
-        // Index the tombstone version with the same interval
-        self.temporal_indexes
-            .insert_edge_version(edge_id, tombstone_id, tombstone_temporal)?;
-
-        // Delete from current storage
-        self.current.delete_edge_direct(edge_id)?;
-
-        Ok(())
-    }
-
-    fn apply_single_write(
-        &self,
-        write: &super::BufferedWrite,
-        commit_timestamp: Timestamp,
-        historical: &mut HistoricalStorage,
-        tombstone_ids: &mut std::vec::IntoIter<u64>,
-        num_deletes: usize,
-    ) -> Result<()> {
-        match write {
-            super::BufferedWrite::CreateNode {
-                node_id,
-                version_id,
-                label,
-                properties,
-                valid_from,
-            } => {
-                self.apply_node_write(
-                    true, // is_create
-                    *node_id,
-                    *version_id,
-                    *label,
-                    properties.clone(),
-                    *valid_from,
-                    commit_timestamp,
-                    historical,
-                )?;
-            }
-            super::BufferedWrite::CreateEdge {
-                edge_id,
-                version_id,
-                source,
-                target,
-                label,
-                properties,
-                valid_from,
-            } => {
-                self.apply_edge_write(
-                    true, // is_create
-                    *edge_id,
-                    *version_id,
-                    *source,
-                    *target,
-                    *label,
-                    properties.clone(),
-                    *valid_from,
-                    commit_timestamp,
-                    historical,
-                )?;
-            }
-            super::BufferedWrite::UpdateNode {
-                node_id,
-                version_id,
-                label,
-                properties,
-                valid_from,
-            } => {
-                self.apply_node_write(
-                    false, // is_create
-                    *node_id,
-                    *version_id,
-                    *label,
-                    properties.clone(),
-                    *valid_from,
-                    commit_timestamp,
-                    historical,
-                )?;
-            }
-            super::BufferedWrite::UpdateEdge {
-                edge_id,
-                version_id,
-                source,
-                target,
-                label,
-                properties,
-                valid_from,
-            } => {
-                self.apply_edge_write(
-                    false, // is_create
-                    *edge_id,
-                    *version_id,
-                    *source,
-                    *target,
-                    *label,
-                    properties.clone(),
-                    *valid_from,
-                    commit_timestamp,
-                    historical,
-                )?;
-            }
-            super::BufferedWrite::DeleteNode {
-                node_id,
-                valid_from,
-            } => {
-                // Use pre-generated tombstone version ID (no lock needed)
-                // CRITICAL: Use proper error handling instead of .expect() to avoid lock poisoning
-                let tombstone_version_id = VersionId::new_unchecked(tombstone_ids.next().ok_or_else(|| {
-                    StorageError::InconsistentState {
-                        reason: format!(
-                            "Tombstone ID exhaustion for DeleteNode: expected {} deletes, iterator depleted at node_id {:?}",
-                            num_deletes, node_id
-                        ),
-                    }
-                })?);
-
-                self.apply_node_delete(
-                    *node_id,
-                    *valid_from,
-                    commit_timestamp,
-                    tombstone_version_id,
-                    historical,
-                )?;
-            }
-            super::BufferedWrite::DeleteEdge {
-                edge_id,
-                valid_from,
-            } => {
-                // Use pre-generated tombstone version ID (no lock needed)
-                // CRITICAL: Use proper error handling instead of .expect() to avoid lock poisoning
-                let tombstone_version_id = VersionId::new_unchecked(tombstone_ids.next().ok_or_else(|| {
-                    StorageError::InconsistentState {
-                        reason: format!(
-                            "Tombstone ID exhaustion for DeleteEdge: expected {} deletes, iterator depleted at edge_id {:?}",
-                            num_deletes, edge_id
-                        ),
-                    }
-                })?);
-
-                self.apply_edge_delete(
-                    *edge_id,
-                    *valid_from,
-                    commit_timestamp,
-                    tombstone_version_id,
-                    historical,
-                )?;
-            }
-        }
-        Ok(())
-    }
-
-    fn apply_changes(&self, commit_timestamp: Timestamp) -> Result<()> {
-        // Create temporal interval for all operations in this transaction.
-        // All operations in a transaction share the same commit_timestamp, which is
-        // semantically correct for transaction_time (when data was recorded).
-        // However, this means the auto-closing logic in add_node_version won't work
-        // for multiple updates in the same transaction (since new_tx_time == prev_tx_time).
-        // Therefore, we explicitly close previous versions before adding new ones.
-        let _temporal = BiTemporalInterval::current(commit_timestamp);
-
-        // Acquire lock on historical storage once before processing all operations.
-        // TemporalIndexes uses DashMap internally, so no outer lock needed.
-        let mut historical = self.historical.write_or_err()?;
-
-        // Pre-generate all tombstone version IDs at once to reduce lock contention
-        // on the ID generator. Count delete operations and generate IDs in batch.
-        let num_deletes = self
-            .buffer
-            .operations()
-            .iter()
-            .filter(|op| {
-                matches!(
-                    op,
-                    super::BufferedWrite::DeleteNode { .. }
-                        | super::BufferedWrite::DeleteEdge { .. }
-                )
-            })
-            .count();
-
-        let mut tombstone_ids = if num_deletes > 0 {
-            let ids: Result<Vec<u64>> = {
-                let id_gen = self.version_id_gen.lock_or_err()?;
-                (0..num_deletes)
-                    .map(|_| id_gen.next().map_err(Into::into))
-                    .collect()
-            };
-            ids?.into_iter()
-        } else {
-            Vec::new().into_iter()
-        };
-
-        for write in self.buffer.operations() {
-            self.apply_single_write(
-                write,
-                commit_timestamp,
-                &mut historical,
-                &mut tombstone_ids,
-                num_deletes,
-            )?;
-        }
-
-        // Safety check: verify all pre-generated tombstone IDs were consumed
-        // This detects count mismatches in development builds
-        debug_assert!(
-            tombstone_ids.next().is_none(),
-            "Tombstone ID surplus: expected {} deletes, but iterator has remaining IDs",
-            num_deletes
-        );
-
-        // Release historical write lock before rebuilding adjacency to avoid holding multiple
-        // exclusive locks simultaneously. TemporalIndexes uses fine-grained DashMap locking
-        // internally, so no explicit lock release needed (automatically released after inserts).
-        drop(historical);
-
-        // Rebuild adjacency indexes only if edge structure changed
-        // With incremental adjacency, edges are immediately visible (no rebuild needed).
-        // However, we compact after transactions with edge operations to maintain
-        // read performance by moving delta edges to frozen CSR.
-        if self.buffer.has_edge_operations() {
-            self.current.compact_adjacency();
-        }
-
-        Ok(())
-    }
-}
-
-impl ReadOps for WriteTransaction {
-    fn get_node(&self, id: NodeId) -> Result<Node> {
-        // Read-your-writes: check write buffer first
-        if let Some(buffered) = self.buffer.get_node_write(id) {
-            match buffered {
-                super::BufferedWrite::CreateNode {
-                    node_id,
-                    label,
-                    properties,
-                    version_id,
-                    ..
-                } => {
-                    // Return the buffered node
-                    return Ok(Node::with_metadata(
-                        *node_id,
-                        *label,
-                        properties.clone(),
-                        *version_id,
-                        VersionMetadata {
-                            created_by_tx: self.tx_id,
-                            commit_timestamp: None, // Not yet committed
-                        },
-                    ));
-                }
-                super::BufferedWrite::UpdateNode {
-                    node_id,
-                    label,
-                    properties,
-                    version_id,
-                    ..
-                } => {
-                    // Return the updated node
-                    return Ok(Node::with_metadata(
-                        *node_id,
-                        *label,
-                        properties.clone(),
-                        *version_id,
-                        VersionMetadata {
-                            created_by_tx: self.tx_id,
-                            commit_timestamp: None,
-                        },
-                    ));
-                }
-                super::BufferedWrite::DeleteNode { .. } => {
-                    // Node has been deleted in this transaction
-                    return Err(StorageError::NodeNotFound(id).into());
-                }
-                _ => {} // Not a node operation
-            }
-        }
-
-        // Fall back to snapshot-isolated read from storage
-        let node = self.current.get_node(id)?;
-
-        // Check if this version is visible in our snapshot
-        if !self
-            .visibility_manager
-            .is_visible(&self.snapshot, node.metadata.created_by_tx)
-        {
-            // Version not visible - return NodeNotFound
-            return Err(StorageError::NodeNotFound(id).into());
-        }
-
-        Ok(node)
-    }
-
-    fn get_edge(&self, id: EdgeId) -> Result<Edge> {
-        // Read-your-writes: check write buffer first
-        if let Some(buffered) = self.buffer.get_edge_write(id) {
-            match buffered {
-                super::BufferedWrite::CreateEdge {
-                    edge_id,
-                    source,
-                    target,
-                    label,
-                    properties,
-                    version_id,
-                    ..
-                } => {
-                    // Return the buffered edge
-                    return Ok(Edge::with_metadata(
-                        *edge_id,
-                        *label,
-                        *source,
-                        *target,
-                        properties.clone(),
-                        *version_id,
-                        VersionMetadata {
-                            created_by_tx: self.tx_id,
-                            commit_timestamp: None,
-                        },
-                    ));
-                }
-                super::BufferedWrite::UpdateEdge {
-                    edge_id,
-                    source,
-                    target,
-                    label,
-                    properties,
-                    version_id,
-                    ..
-                } => {
-                    // Return the updated edge
-                    return Ok(Edge::with_metadata(
-                        *edge_id,
-                        *label,
-                        *source,
-                        *target,
-                        properties.clone(),
-                        *version_id,
-                        VersionMetadata {
-                            created_by_tx: self.tx_id,
-                            commit_timestamp: None,
-                        },
-                    ));
-                }
-                super::BufferedWrite::DeleteEdge { .. } => {
-                    // Edge has been deleted in this transaction
-                    return Err(StorageError::EdgeNotFound(id).into());
-                }
-                _ => {} // Not an edge operation
-            }
-        }
-
-        // Fall back to snapshot-isolated read from storage
-        let edge = self.current.get_edge(id)?;
-
-        // Check if this version is visible in our snapshot
-        if !self
-            .visibility_manager
-            .is_visible(&self.snapshot, edge.metadata.created_by_tx)
-        {
-            // Version not visible - return EdgeNotFound
-            return Err(StorageError::EdgeNotFound(id).into());
-        }
-
-        Ok(edge)
-    }
-
-    fn get_outgoing_edges(&self, node_id: NodeId) -> Vec<EdgeId> {
-        self.current.get_outgoing_edges(node_id)
-    }
-
-    fn get_incoming_edges(&self, node_id: NodeId) -> Vec<EdgeId> {
-        self.current.get_incoming_edges(node_id)
-    }
-
-    fn get_outgoing_edges_with_label(&self, node_id: NodeId, label: &str) -> Vec<EdgeId> {
-        self.current.get_outgoing_edges_with_label(node_id, label)
-    }
-
-    fn node_count(&self) -> usize {
-        self.current.node_count()
-    }
-
-    fn edge_count(&self) -> usize {
-        self.current.edge_count()
-    }
-}
-
-impl WriteOps for WriteTransaction {
-    fn create_node_with_valid_time(
-        &mut self,
-        label: &str,
-        properties: PropertyMap,
-        valid_from: Option<Timestamp>,
-    ) -> Result<NodeId> {
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Generate IDs
-        let node_id = NodeId::new_unchecked(self.node_id_gen.lock_or_err()?.next()?);
-        let version_id = VersionId::new_unchecked(self.version_id_gen.lock_or_err()?.next()?);
-        let label_interned = GLOBAL_INTERNER.intern(label)?;
-
-        // Get timestamp: use provided valid_from or default to transaction start time
-        let timestamp = self.start_timestamp;
-        let valid_from = valid_from.unwrap_or(timestamp);
-
-        // Validate valid_from is not too far in future
-        self.validate_valid_from_future(valid_from)?;
-
-        // Buffer the write
-        self.buffer.add(super::BufferedWrite::CreateNode {
-            node_id,
-            version_id,
-            label: label_interned,
-            properties,
-            valid_from,
-        })?;
-
-        Ok(node_id)
-    }
-
-    fn create_edge_with_valid_time(
-        &mut self,
-        source: NodeId,
-        target: NodeId,
-        label: &str,
-        properties: PropertyMap,
-        valid_from: Option<Timestamp>,
-    ) -> Result<EdgeId> {
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Generate IDs
-        let edge_id = EdgeId::new_unchecked(self.edge_id_gen.lock_or_err()?.next()?);
-        let version_id = VersionId::new_unchecked(self.version_id_gen.lock_or_err()?.next()?);
-        let label_interned = GLOBAL_INTERNER.intern(label)?;
-
-        // Get timestamp: use provided valid_from or default to transaction start time
-        let timestamp = self.start_timestamp;
-        let valid_from = valid_from.unwrap_or(timestamp);
-
-        // Buffer the write
-        self.buffer.add(super::BufferedWrite::CreateEdge {
-            edge_id,
-            version_id,
-            source,
-            target,
-            label: label_interned,
-            properties,
-            valid_from,
-        })?;
-
-        Ok(edge_id)
-    }
-
-    fn update_node_with_valid_time(
-        &mut self,
-        node_id: NodeId,
-        properties: PropertyMap,
-        valid_from: Option<Timestamp>,
-    ) -> Result<()> {
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Get current node to preserve label and existing properties
-        let node = self.current.get_node(node_id)?;
-        let version_id = VersionId::new_unchecked(self.version_id_gen.lock_or_err()?.next()?);
-
-        // PATCH semantics: Merge new properties with existing ones
-        // Start with existing properties
-        let mut builder = PropertyMapBuilder::from_map(node.properties.clone());
-
-        // Update/add properties from the incoming map
-        for (key, value) in properties.iter() {
-            builder = builder.insert_by_key(*key, value.clone());
-        }
-
-        // Build the final merged property map
-        let merged_properties = builder.build();
-
-        // Get timestamp: use provided valid_from or default to transaction start time
-        let timestamp = self.start_timestamp;
-        let valid_from = valid_from.unwrap_or(timestamp);
-
-        // Validate valid_from is not too far in future
-        self.validate_valid_from_future(valid_from)?;
-
-        // Validate valid_from is not before entity creation
-        let historical = self.historical.read_or_err()?;
-        if let Some(current_version_id) = historical.get_current_node_version(node_id)
-            && let Some(current_version) = historical.get_node_version(current_version_id)
-        {
-            let creation_time = current_version.temporal.valid_time().start();
-            drop(historical); // Release lock before calling validation
-            self.validate_valid_from_not_before_creation(
-                &format!("node:{}", node_id.as_u64()),
-                creation_time,
-                valid_from,
-            )?;
-        }
-
-        // Buffer the write with merged properties
-        self.buffer.add(super::BufferedWrite::UpdateNode {
-            node_id,
-            version_id,
-            label: node.label,
-            properties: merged_properties,
-            valid_from,
-        })?;
-
-        Ok(())
-    }
-
-    fn update_edge_with_valid_time(
-        &mut self,
-        edge_id: EdgeId,
-        properties: PropertyMap,
-        valid_from: Option<Timestamp>,
-    ) -> Result<()> {
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Get current edge to preserve source, target, label and existing properties
-        let edge = self.current.get_edge(edge_id)?;
-        let version_id = VersionId::new_unchecked(self.version_id_gen.lock_or_err()?.next()?);
-
-        // PATCH semantics: Merge new properties with existing ones
-        // Start with existing properties
-        let mut builder = PropertyMapBuilder::from_map(edge.properties.clone());
-
-        // Update/add properties from the incoming map
-        for (key, value) in properties.iter() {
-            builder = builder.insert_by_key(*key, value.clone());
-        }
-
-        // Build the final merged property map
-        let merged_properties = builder.build();
-
-        // Get timestamp: use provided valid_from or default to transaction start time
-        let timestamp = self.start_timestamp;
-        let valid_from = valid_from.unwrap_or(timestamp);
-
-        // Buffer the write with merged properties
-        self.buffer.add(super::BufferedWrite::UpdateEdge {
-            edge_id,
-            version_id,
-            source: edge.source,
-            target: edge.target,
-            label: edge.label,
-            properties: merged_properties,
-            valid_from,
-        })?;
-
-        Ok(())
-    }
-
-    fn delete_node_with_valid_time(
-        &mut self,
-        node_id: NodeId,
-        valid_from: Option<Timestamp>,
-    ) -> Result<()> {
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Verify node exists and check for vector properties
-        let node = self.current.get_node(node_id)?;
-
-        // If the node being deleted contains vector properties, mark the buffer
-        // to ensure the temporal vector index is notified on commit
-        if !self.buffer.has_vector_operations() && node.properties.contains_vector() {
-            self.buffer.mark_has_vector_operations();
-        }
-
-        // Get timestamp: use provided valid_from or default to transaction start time
-        let timestamp = self.start_timestamp;
-        let valid_from = valid_from.unwrap_or(timestamp);
-
-        // Validate valid_from is not too far in future
-        self.validate_valid_from_future(valid_from)?;
-
-        // Validate valid_from is not before entity creation
-        let historical = self.historical.read_or_err()?;
-        if let Some(current_version_id) = historical.get_current_node_version(node_id)
-            && let Some(current_version) = historical.get_node_version(current_version_id)
-        {
-            let creation_time = current_version.temporal.valid_time().start();
-            drop(historical); // Release lock before calling validation
-            self.validate_valid_from_not_before_creation(
-                &format!("node:{}", node_id.as_u64()),
-                creation_time,
-                valid_from,
-            )?;
-        }
-
-        // Buffer the write
-        self.buffer.add(super::BufferedWrite::DeleteNode {
-            node_id,
-            valid_from,
-        })?;
-
-        Ok(())
-    }
-
-    fn delete_node_cascade(&mut self, node_id: NodeId) -> Result<()> {
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Verify node exists before attempting deletion
-        let _node = self.current.get_node(node_id)?;
-
-        // Collect all edges connected to this node (both outgoing and incoming)
-        // We do this before any deletions to avoid borrowing issues
-        //
-        // LIMITATION: This uses ReadOps methods which currently don't support
-        // read-your-writes semantics for edge traversal. This means edges created
-        // in the same transaction (but not yet committed) won't be found and deleted.
-        // This is consistent with the existing ReadOps behavior but may leave orphaned
-        // edges in same-transaction scenarios. See issue for future improvement.
-        let outgoing_edges = self.get_outgoing_edges(node_id);
-        let incoming_edges = self.get_incoming_edges(node_id);
-
-        // Delete all connected edges first to maintain referential integrity
-        // This prevents orphaned edges that reference a deleted node
-        // Performance: O(degree) where degree is the number of connected edges
-        for edge_id in outgoing_edges.into_iter().chain(incoming_edges) {
-            self.delete_edge(edge_id)?;
-        }
-
-        // Finally, delete the node itself
-        // This is safe now because all edges referencing this node have been removed
-        self.delete_node(node_id)?;
-
-        Ok(())
-    }
-
-    fn delete_edge_with_valid_time(
-        &mut self,
-        edge_id: EdgeId,
-        valid_from: Option<Timestamp>,
-    ) -> Result<()> {
-        // Check transaction state
-        if self.state != TxState::Active {
-            return Err(TransactionError::InvalidState {
-                current: format!("{:?}", self.state),
-                expected: "Active".to_string(),
-            }
-            .into());
-        }
-
-        // Verify edge exists and check for vector properties
-        let edge = self.current.get_edge(edge_id)?;
-
-        // If the edge being deleted contains vector properties, mark the buffer
-        // to ensure the temporal vector index is notified on commit
-        if !self.buffer.has_vector_operations() && edge.properties.contains_vector() {
-            self.buffer.mark_has_vector_operations();
-        }
-
-        // Get timestamp: use provided valid_from or default to transaction start time
-        let timestamp = self.start_timestamp;
-        let valid_from = valid_from.unwrap_or(timestamp);
-
-        // Buffer the write
-        self.buffer.add(super::BufferedWrite::DeleteEdge {
-            edge_id,
-            valid_from,
-        })?;
-
-        Ok(())
-    }
-}
-
-impl Drop for WriteTransaction {
-    fn drop(&mut self) {
-        // Auto-rollback if not committed
-        if self.state == TxState::Active {
-            self.buffer.clear();
-            // Register abort with visibility manager
-            self.visibility_manager.register_abort(self.tx_id);
-            self.state = TxState::Aborted;
-        }
-    }
-}
-
-#[cfg(test)]
 mod tombstone_tests {
     use super::*;
-    use crate::api::transaction::TxIdGenerator;
-    use crate::storage::wal::concurrent_system::ConcurrentWalSystemConfig;
-    use tempfile::TempDir;
 
     fn create_test_write_tx() -> (WriteTransaction, TempDir) {
         let current = Arc::new(CurrentStorage::new());
@@ -1896,7 +60,8 @@ mod tombstone_tests {
         };
 
         // Try to apply it
-        let result = tx.apply_single_write(
+        let result = crate::api::transaction::write::apply::apply_single_write(
+            &tx,
             &op,
             commit_ts,
             &mut historical,
@@ -1914,13 +79,9 @@ mod tombstone_tests {
     }
 }
 
-#[cfg(test)]
-mod tests {
+mod general_tests {
     use super::*;
-    use crate::api::transaction::TxIdGenerator;
     use crate::core::property::PropertyMapBuilder;
-    use crate::storage::wal::concurrent_system::ConcurrentWalSystemConfig;
-    use tempfile::TempDir;
 
     fn create_test_write_tx() -> (WriteTransaction, TempDir) {
         let current = Arc::new(CurrentStorage::new());
@@ -3403,11 +1564,6 @@ mod tests {
     }
 }
 
-/// Tests for write-write conflict detection (Issue #8).
-///
-/// These tests verify that concurrent transactions updating the same entity
-/// will properly detect conflicts and fail with SerializationFailure.
-#[cfg(test)]
 mod conflict_detection_tests {
     use super::*;
     use crate::api::transaction::TxIdGenerator;
@@ -4323,10 +2479,10 @@ mod conflict_detection_tests {
     }
 }
 
-#[cfg(test)]
 mod clock_skew_tests {
     use super::*;
     use crate::api::transaction::TxIdGenerator;
+    use crate::core::property::PropertyMapBuilder;
     use crate::storage::wal::concurrent_system::ConcurrentWalSystemConfig;
     use tempfile::TempDir;
 
@@ -4461,12 +2617,6 @@ mod clock_skew_tests {
     }
 }
 
-/// Tests for concurrent commit timestamp ordering (Issue #10).
-///
-/// These tests verify that transaction timestamps are assigned in the same
-/// order as their commits become durable. This is critical for bi-temporal
-/// correctness - transaction time must be monotonic with actual commit order.
-#[cfg(test)]
 mod timestamp_ordering_tests {
     use super::*;
     use crate::api::transaction::TxIdGenerator;
@@ -4727,14 +2877,10 @@ mod timestamp_ordering_tests {
     }
 }
 
-// =========================================================================
-// Phase 5: Bi-Temporal Validation Tests
-// =========================================================================
-
-#[cfg(test)]
 mod bitemporal_validation_tests {
     use super::*;
     use crate::api::transaction::types::TxIdGenerator;
+    use crate::core::property::PropertyMap;
     use crate::storage::wal::concurrent_system::ConcurrentWalSystemConfig;
     use tempfile::TempDir;
